Move letterbox into yolort.models.transform
